# api/openapi.yaml
openapi: 3.0.3
info:
  title: Cars API
  version: 1.0.0
  description: CRUD API for cars backed by SQLite.
servers:
  - url: http://localhost:8080
paths:
  /api/cars:
    get:
      summary: List cars
      operationId: listCars
      responses:
        '200':
          description: Cars list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendCarsSuccess'
    post:
      summary: Create car
      operationId: createCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarInput'
      responses:
        '201':
          description: Car created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendCarSuccess'
        '400':
          description: Validation or payload error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendFail'
  /api/cars/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          format: int64
    get:
      summary: Get car by ID
      operationId: getCarByID
      responses:
        '200':
          description: Car found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendCarSuccess'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendError'
    put:
      summary: Update car
      operationId: updateCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarInput'
      responses:
        '200':
          description: Car updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendCarSuccess'
        '400':
          description: Validation or payload error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendFail'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendError'
    delete:
      summary: Delete car
      operationId: deleteCar
      responses:
        '200':
          description: Car deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendDeleteSuccess'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JSendError'
components:
  schemas:
    Car:
      type: object
      required: [id, inventory_id, make, model, year, color, vin]
      properties:
        id:
          type: integer
          format: int64
        inventory_id:
          type: integer
          format: int64
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        vin:
          type: string
    CarInput:
      type: object
      required: [inventory_id, make, model, year, color, vin]
      properties:
        inventory_id:
          type: integer
          format: int64
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        vin:
          type: string
    JSendCarSuccess:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          $ref: '#/components/schemas/Car'
    JSendCarsSuccess:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Car'
    JSendDeleteSuccess:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          enum: [success]
        data:
          type: object
          required: [deleted]
          properties:
            deleted:
              type: boolean
    JSendFail:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [fail]
        message:
          type: string
    JSendError:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [error]
        message:
          type: string